move all to proper directory
